+++ hotel-management/src/components/Nav.tsx
@@ -1,5 +1,7 @@
 import { NavLink } from 'react-router-dom';
 import styled from 'styled-components';
+
+// Components
 import { MdSpaceDashboard } from 'react-icons/md';
 import { HiUsers } from 'react-icons/hi2';
 import { BsHouseDoorFill } from 'react-icons/bs';
