+++ hotel-management/src/globals/interfaces.ts
@@ -9,3 +9,13 @@ export interface IButton {
   icon?: JSX.Element;
   onClick?: () => void;
 }
+
+export interface ITable {
+  $columns?: string;
+  children: React.ReactNode;
+}
+
+export interface ITableBody<T> {
+  data?: T[];
+  render?: (value: T) => JSX.Element;
+}
